Move error tests to derive.rs
